OpenCV 4 compatibility and default. OpenCV 3 compat likely broken now
Implementation notes beyond what this commit adds to the CHANGELOG:

* `drawAxis()` was replaced by `drawFrameAxes()`.
  The former is available only until OpenCV <= 4.5,
  while the latter is available for all 4.x.
  * https://github.com/opencv/opencv_contrib/commit/56d492cfecb7aadaca79a9494a33ad1c12798e5e
  * https://github.com/opencv/opencv_contrib/commit/0f030dd002e0daaa9f5c4ba705796cb56b8e9bff

+++ opencv/src/OpenCV/Internal/ImgProc/StructuralAnalysis.hs
@@ -114,7 +114,7 @@ approxPolyDP curve epsilon isClosed
             approxSizePtr
             approxPtrPtr
             (toCDouble epsilon)
-            (fromIntegral (fromBool isClosed))
+            (fromBool isClosed)
 
 -- | Internal class used to overload the 'approxPolyDP' depth.
 class ( FromPtr      (Point   2 depth)
@@ -322,7 +322,7 @@ convexHull points clockwise
           convexHull_internal
             (fromIntegral $ V.length points)
             pointsPtr
-            (fromIntegral (fromBool clockwise))
+            (fromBool clockwise)
             hullPointsPtrPtr
             hullSizePtr
   where
@@ -449,7 +449,7 @@ convexHullIndices points clockwise
           convexHullIndices_internal
             (fromIntegral $ V.length points)
             pointsPtr
-            (fromIntegral (fromBool clockwise))
+            (fromBool clockwise)
             hullIndicesPtrPtr
             hullSizePtr
   where
